fixed mistakes
