+++ Task05/CoarseTests/Delete.cs
@@ -1,15 +1,14 @@
 using NUnit.Framework;
-using Task05;
 
-namespace Task5.FineTests
+namespace Task05.CoarseTests
 {
     [TestFixture]
     public class DeleteTests
     {
         [TestCase(0, 0)]
-        public void Delete_RootInFineTree_ReturnEmptyTree(int key, int value)
+        public void Delete_RootInCoarseTree_ReturnEmptyTree(int key, int value)
         {
-            var tree = new BSTFineGrained<int>();
+            var tree = new BSTCoarseGrained<int>();
             tree.Insert(key, value);
 
             tree.Delete(key);
@@ -18,9 +17,9 @@ namespace Task5.FineTests
         }
 
         [TestCase(0, 0)]
-        public void Delete_NonExistentNodeInFineTree_ReturnEmptyTree(int key, int value)
+        public void Delete_NonExistentNodeInCoarseTree_ReturnEmptyTree(int key, int value)
         {
-            var tree = new BSTFineGrained<int>();
+            var tree = new BSTCoarseGrained<int>();
             tree.Insert(key, value);
 
             Assert.DoesNotThrow(() => tree.Delete(value + 1));
@@ -31,10 +30,10 @@ namespace Task5.FineTests
         }
 
         [TestCase(5, 5, 3, 3, 1, 1)]
-        public void Delete_RootInFineTree_SetNewRootFromLeft(int rootKey, int rootValue, int leftKey,
+        public void Delete_RootInCoarseTree_SetNewRootFromLeft(int rootKey, int rootValue, int leftKey,
             int leftValue, int leftOfLeftKey, int leftOfLeftValue)
         {
-            var tree = new BSTFineGrained<int>();
+            var tree = new BSTCoarseGrained<int>();
             tree.Insert(rootKey, rootValue);
             tree.Insert(leftKey, leftValue);
             tree.Insert(leftOfLeftKey, leftOfLeftValue);
@@ -51,10 +50,10 @@ namespace Task5.FineTests
         }
 
         [TestCase(5, 5, 7, 7, 9, 9)]
-        public void Delete_RootInFineTree_SetNewRootFromRight(int rootKey, int rootValue, int rightKey,
+        public void Delete_RootInCoarseTree_SetNewRootFromRight(int rootKey, int rootValue, int rightKey,
             int rightValue, int rightOfRightKey, int rightOfRightValue)
         {
-            var tree = new BSTFineGrained<int>();
+            var tree = new BSTCoarseGrained<int>();
             tree.Insert(rootKey, rootValue);
             tree.Insert(rightKey, rightValue);
             tree.Insert(rightOfRightKey, rightOfRightValue);
@@ -71,11 +70,11 @@ namespace Task5.FineTests
         }
 
         [TestCase(5, 5, 2, 2, 4, 4, 3, 3)]
-        public void Delete_RootInFineTree_UpdateRootFromLeftNeighbor(int rootKey, int rootValue, int leftKey,
+        public void Delete_RootInCoarseTree_UpdateRootFromLeftNeighbor(int rootKey, int rootValue, int leftKey,
             int leftValue, int rightOfLeftKey, int rightOfLeftValue, int leftOfRightOfLeftKey,
             int leftOfRightOfLeftValue)
         {
-            var tree = new BSTFineGrained<int>();
+            var tree = new BSTCoarseGrained<int>();
             tree.Insert(rootKey, rootValue);
             tree.Insert(leftKey, leftValue);
             tree.Insert(rightOfLeftKey, rightOfLeftValue);
@@ -97,11 +96,11 @@ namespace Task5.FineTests
         }
 
         [TestCase(5, 5, 8, 8, 6, 6, 7, 7)]
-        public void Delete_RootInFineTree_UpdateRootFromRightNeighbor(int rootKey, int rootValue, int rightKey,
+        public void Delete_RootInCoarseTree_UpdateRootFromRightNeighbor(int rootKey, int rootValue, int rightKey,
             int rightValue, int leftOrRightKey, int leftOfRightValue, int rightOfLeftOfOfRightKey,
             int rightOfLeftOfOfRightValue)
         {
-            var tree = new BSTFineGrained<int>();
+            var tree = new BSTCoarseGrained<int>();
             tree.Insert(rootKey, rootValue);
             tree.Insert(rightKey, rightValue);
             tree.Insert(leftOrRightKey, leftOfRightValue);
@@ -123,11 +122,11 @@ namespace Task5.FineTests
         }
 
         [TestCase(5, 5, 8, 8, 7, 7, 6, 6)]
-        public void Delete_InnerNodeInFineTree_UpdateFromRight(int rootKey, int rootValue, int rightKey,
+        public void Delete_InnerNodeInCoarseTree_UpdateFromRight(int rootKey, int rootValue, int rightKey,
             int rightValue, int leftOrRightKey, int leftOfRightValue, int leftOfLeftOfRightKey,
             int leftOfLeftOfRightValue)
         {
-            var tree = new BSTFineGrained<int>();
+            var tree = new BSTCoarseGrained<int>();
             tree.Insert(rootKey, rootValue);
             tree.Insert(rightKey, rightValue);
             tree.Insert(leftOrRightKey, leftOfRightValue);
@@ -149,11 +148,11 @@ namespace Task5.FineTests
         }
 
         [TestCase(5, 5, 1, 1, 3, 3, 4, 4)]
-        public void Delete_InnerNodeInFineTree_UpdateFromLeft(int rootKey, int rootValue, int leftKey,
+        public void Delete_InnerNodeInCoarseTree_UpdateFromLeft(int rootKey, int rootValue, int leftKey,
             int leftValue, int rightOfLeftKey, int rightOfLeftValue, int rightOfRightOfLeftKey,
             int rightOfRightOfLeftValue)
         {
-            var tree = new BSTFineGrained<int>();
+            var tree = new BSTCoarseGrained<int>();
             tree.Insert(rootKey, rootValue);
             tree.Insert(leftKey, leftValue);
             tree.Insert(rightOfLeftKey, rightOfLeftValue);
@@ -175,9 +174,9 @@ namespace Task5.FineTests
         }
 
         [TestCase(1, 1, 2, 2)]
-        public void Delete_RightLeafInFineTree_UpdateTree(int rootKey, int rootValue, int rightKey, int rightValue)
+        public void Delete_RightLeafInCoarseTree_UpdateTree(int rootKey, int rootValue, int rightKey, int rightValue)
         {
-            var tree = new BSTFineGrained<int>();
+            var tree = new BSTCoarseGrained<int>();
             tree.Insert(rootKey, rootValue);
             tree.Insert(rightKey, rightValue);
 
@@ -189,9 +188,9 @@ namespace Task5.FineTests
         }
 
         [TestCase(2, 2, 1, 1)]
-        public void Delete_LeftLeafInFineTree_UpdateTree(int rootKey, int rootValue, int leftKey, int leftValue)
+        public void Delete_LeftLeafInCoarseTree_UpdateTree(int rootKey, int rootValue, int leftKey, int leftValue)
         {
-            var tree = new BSTFineGrained<int>();
+            var tree = new BSTCoarseGrained<int>();
             tree.Insert(rootKey, rootValue);
             tree.Insert(leftKey, leftValue);
