+++ Task05/Node.cs
