+++ Task05/Program.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using Task5;
 
 namespace Task05
 {
